Update README.md
I2C is a two-wire serial communication protocol used for communication between integrated circuits. In this project, I2C is used to communicate with both the SHT31 sensor and the SSD1306 OLED display. The SHT31 sensor is accessed through I2C for reading temperature and humidity data. The relevant code is: